Broke functions out into files... Not sure if this will work
